+++ HISTORY.rst
@@ -10,6 +10,8 @@ Local tuner
 
 - Add a plot which shows the optima predicted by the tuner across the
   iterations (#172). This can be useful to gauge convergence.
+- Add a plot which shows the estimated Elo (+ confidence interval) of the
+  predicted optima (#176).
 - Tuner saves optima and their Elo performance (including standard deviation)
   to disk now (#171).